added factory class
