+++ batch7crud-roll4/src/main/java/com/lftechnology/batch7crud/validator/EmployeeValidator.java
@@ -2,6 +2,7 @@ package com.lftechnology.batch7crud.validator;
 
 import static com.lftechnology.batch7crud.constant.ParamConstants.*;
 
+import com.lftechnology.batch7crud.exception.ValidationException;
 import com.lftechnology.batch7crud.model.Employee;
 
 import java.util.HashMap;
@@ -13,32 +14,18 @@ import java.util.Map;
 public class EmployeeValidator implements Validator<Employee> {
 
   @Override
-  public Employee createObject(Map<String, String> input) {
-    String firstName = input.get(PARAM_FIRST_NAME).trim();
-    String lastName = input.get(PARAM_LAST_NAME).trim();
-    String station = input.get(PARAM_STATION).trim();
-
-    Employee employee = new Employee();
-    employee.setFirstName(firstName);
-    employee.setLastName(lastName);
-    employee.setStation(station);
-
-    return employee;
-  }
-
-  @Override
-  public Map<String, String> validate(Employee e) {
+  public void validate(Employee e) throws ValidationException {
     Map<String, String> errors = new HashMap<String, String>();
 
     String firstName = e.getFirstName().trim();
     String lastName = e.getLastName().trim();
     String station = e.getStation().trim();
 
-    if (isNullOrEmpty(firstName) || !isString(firstName)) {
+    if (isNullOrEmpty(firstName) || !isAlphabet(firstName)) {
       errors.put(PARAM_FIRST_NAME, "Check first name");
     }
 
-    if (isNullOrEmpty(lastName) || !isString(lastName)) {
+    if (isNullOrEmpty(lastName) || !isAlphabet(lastName)) {
       errors.put(PARAM_LAST_NAME, "Check Last Name");
     }
 
@@ -46,14 +33,16 @@ public class EmployeeValidator implements Validator<Employee> {
       errors.put(PARAM_STATION, "Check Station");
     }
 
-    return errors;
+    if(!errors.isEmpty()) {
+      throw new ValidationException(errors);
+    }
   }
 
   private boolean isNullOrEmpty(String argument) {
-    return "".equals(argument);
+    return "".equals(argument) || argument == null;
   }
 
-  private boolean isString(String argument) {
+  private boolean isAlphabet(String argument) {
     return argument.matches("[A-Za-z]*");
   }
