+++ batch7crud-roll4/src/main/java/com/lftechnology/batch7crud/validator/Validator.java
@@ -1,12 +1,12 @@
 package com.lftechnology.batch7crud.validator;
 
+import com.lftechnology.batch7crud.exception.ValidationException;
+
 import java.util.Map;
 
 /**
  * Created by pratishshr on 1/25/16.
  */
 public interface Validator<T> {
-  public T createObject(Map<String, String> input);
-
-  public Map<String, String> validate(T entity);
+  public void validate(T entity) throws ValidationException;
 }
